feat(theme): extract theme system to separate crate
Create jolt-theme crate with:
- Color types and ThemeColors struct (platform-agnostic)
- Theme loading: builtin (embedded TOML) and user themes
- Theme validation with WCAG contrast checking
- iTerm2 color scheme import functionality
- Scheme list caching

CLI now depends on jolt-theme and provides thin wrappers
for path resolution (themes_dir, cache_dir).
