+++ cli/Cargo.toml
@@ -11,6 +11,7 @@ categories = ["command-line-utilities"]
 
 [dependencies]
 protocol = { workspace = true }
+jolt-theme = { workspace = true }
 ratatui = { workspace = true }
 crossterm = { workspace = true }
 tokio = { workspace = true }
