+++ crates/theme/src/themes/catppuccin.toml
@@ -1,0 +1,33 @@
+name = "Catppuccin"
+
+[dark]
+bg = "#1e1e2e"
+dialog_bg = "#313244"
+fg = "#cdd6f4"
+accent = "#89b4fa"
+accent_secondary = "#cba6f7"
+highlight = "#f9e2af"
+muted = "#9399b2"
+success = "#a6e3a1"
+warning = "#fab387"
+danger = "#f38ba8"
+border = "#45475a"
+selection_bg = "#585b70"
+selection_fg = "#cdd6f4"
+graph_line = "#89b4fa"
+
+[light]
+bg = "#eff1f5"
+dialog_bg = "#e6e9ef"
+fg = "#4c4f69"
+accent = "#1557d2"
+accent_secondary = "#8839ef"
+highlight = "#825a0a"
+muted = "#5c5f77"
+success = "#227815"
+warning = "#b44600"
+danger = "#b40a2d"
+border = "#bcc0cc"
+selection_bg = "#9ca0b0"
+selection_fg = "#303244"
+graph_line = "#1557d2"
