+++ crates/theme/src/themes/default.toml
@@ -1,0 +1,33 @@
+name = "Default"
+
+[dark]
+bg = "#16161e"
+dialog_bg = "#23232d"
+fg = "#e6e6f0"
+accent = "#8ab4f8"
+accent_secondary = "#bb86fc"
+highlight = "#ffcb6b"
+muted = "#80808c"
+success = "#81c784"
+warning = "#ffb74d"
+danger = "#ef5350"
+border = "#3c3c50"
+selection_bg = "#323246"
+selection_fg = "#ffffff"
+graph_line = "#8ab4f8"
+
+[light]
+bg = "#fafafc"
+dialog_bg = "#ffffff"
+fg = "#1e1e28"
+accent = "#145faa"
+accent_secondary = "#7b1fa2"
+highlight = "#a06400"
+muted = "#64646e"
+success = "#2e7d32"
+warning = "#af5000"
+danger = "#d32f2f"
+border = "#c8c8d2"
+selection_bg = "#dce6f5"
+selection_fg = "#000000"
+graph_line = "#145faa"
