+++ crates/theme/src/themes/dracula.toml
@@ -1,0 +1,33 @@
+name = "Dracula"
+
+[dark]
+bg = "#282a36"
+dialog_bg = "#44475a"
+fg = "#f8f8f2"
+accent = "#caa9fa"
+accent_secondary = "#ff79c6"
+highlight = "#f1fa8c"
+muted = "#8c9bc8"
+success = "#50fa7b"
+warning = "#ffb86c"
+danger = "#ff5555"
+border = "#44475a"
+selection_bg = "#44475a"
+selection_fg = "#f8f8f2"
+graph_line = "#8be9fd"
+
+[light]
+bg = "#f8f8f2"
+dialog_bg = "#ffffff"
+fg = "#282a36"
+accent = "#6e469c"
+accent_secondary = "#a03273"
+highlight = "#78691e"
+muted = "#505f8c"
+success = "#198237"
+warning = "#a06428"
+danger = "#b42323"
+border = "#c8c8d2"
+selection_bg = "#dcdce6"
+selection_fg = "#282a36"
+graph_line = "#328296"
