+++ crates/theme/src/themes/gruvbox.toml
@@ -1,0 +1,33 @@
+name = "Gruvbox"
+
+[dark]
+bg = "#282828"
+dialog_bg = "#3c3836"
+fg = "#ebdbb2"
+accent = "#96b9ac"
+accent_secondary = "#d3869b"
+highlight = "#fabd2f"
+muted = "#a8998a"
+success = "#b8bb26"
+warning = "#fe8019"
+danger = "#ff6450"
+border = "#504945"
+selection_bg = "#665c54"
+selection_fg = "#ebdbb2"
+graph_line = "#8ec07c"
+
+[light]
+bg = "#fbf1c7"
+dialog_bg = "#f2e5bc"
+fg = "#3c3836"
+accent = "#32696c"
+accent_secondary = "#91466e"
+highlight = "#825a0a"
+muted = "#6e5f50"
+success = "#6e6e0f"
+warning = "#aa460a"
+danger = "#af1914"
+border = "#d5c4a1"
+selection_bg = "#bdae93"
+selection_fg = "#282422"
+graph_line = "#4b784e"
