+++ crates/theme/src/themes/monokai.toml
@@ -1,0 +1,33 @@
+name = "Monokai"
+
+[dark]
+bg = "#272822"
+dialog_bg = "#35362f"
+fg = "#f8f8f2"
+accent = "#66d9ef"
+accent_secondary = "#ae81ff"
+highlight = "#e6db74"
+muted = "#9b9682"
+success = "#a6e22e"
+warning = "#fd971f"
+danger = "#ff5591"
+border = "#49483e"
+selection_bg = "#49483e"
+selection_fg = "#f8f8f2"
+graph_line = "#66d9ef"
+
+[light]
+bg = "#fdfdfd"
+dialog_bg = "#ffffff"
+fg = "#272822"
+accent = "#0f6e7d"
+accent_secondary = "#6441a0"
+highlight = "#786e19"
+muted = "#5a5746"
+success = "#4b7828"
+warning = "#a05f0a"
+danger = "#a00f4b"
+border = "#c8c8be"
+selection_bg = "#e6e6dc"
+selection_fg = "#272822"
+graph_line = "#0f6e7d"
