+++ crates/theme/src/themes/nord.toml
@@ -1,0 +1,33 @@
+name = "Nord"
+
+[dark]
+bg = "#2e3440"
+dialog_bg = "#3b4252"
+fg = "#d8dee9"
+accent = "#88c0d0"
+accent_secondary = "#c8a5c3"
+highlight = "#ebcb8b"
+muted = "#96a0b4"
+success = "#a3be8c"
+warning = "#dc9b87"
+danger = "#e18c96"
+border = "#434c5e"
+selection_bg = "#4c566a"
+selection_fg = "#eceff4"
+graph_line = "#81a1c1"
+
+[light]
+bg = "#eceff4"
+dialog_bg = "#e5e9f0"
+fg = "#3b4252"
+accent = "#46648c"
+accent_secondary = "#825f7d"
+highlight = "#786437"
+muted = "#4c566a"
+success = "#4b693a"
+warning = "#8c5037"
+danger = "#a0414b"
+border = "#d8dee9"
+selection_bg = "#c8d0dc"
+selection_fg = "#2e3440"
+graph_line = "#46648c"
