+++ crates/theme/src/themes/one-dark.toml
@@ -1,0 +1,33 @@
+name = "One Dark"
+
+[dark]
+bg = "#282c34"
+dialog_bg = "#353b45"
+fg = "#abb2bf"
+accent = "#61afef"
+accent_secondary = "#c678dd"
+highlight = "#e5c07b"
+muted = "#9199aa"
+success = "#98c379"
+warning = "#d19a66"
+danger = "#eb828c"
+border = "#636d83"
+selection_bg = "#3e4451"
+selection_fg = "#abb2bf"
+graph_line = "#56b6c2"
+
+[light]
+bg = "#fafafa"
+dialog_bg = "#ffffff"
+fg = "#383a42"
+accent = "#2d5fc3"
+accent_secondary = "#a626a4"
+highlight = "#916400"
+muted = "#6e6e76"
+success = "#377d37"
+warning = "#825a00"
+danger = "#c33c32"
+border = "#c8c8c8"
+selection_bg = "#e6e6e6"
+selection_fg = "#383a42"
+graph_line = "#006996"
