+++ crates/theme/src/themes/rose-pine.toml
@@ -1,0 +1,33 @@
+name = "Rosé Pine"
+
+[dark]
+bg = "#191724"
+dialog_bg = "#1f1d2e"
+fg = "#e0def4"
+accent = "#559bb4"
+accent_secondary = "#c4a7e7"
+highlight = "#f6c177"
+muted = "#918caa"
+success = "#9ccfd8"
+warning = "#f6c177"
+danger = "#eb6f92"
+border = "#26233a"
+selection_bg = "#403d52"
+selection_fg = "#e0def4"
+graph_line = "#ebbcba"
+
+[light]
+bg = "#faf4ed"
+dialog_bg = "#fffaf3"
+fg = "#575279"
+accent = "#1e556e"
+accent_secondary = "#735f8c"
+highlight = "#8c5f14"
+muted = "#6e697d"
+success = "#3c737d"
+warning = "#8c5f14"
+danger = "#964b5f"
+border = "#f2e9de"
+selection_bg = "#d2c8c3"
+selection_fg = "#3c3755"
+graph_line = "#af645f"
