+++ crates/theme/src/themes/solarized.toml
@@ -1,0 +1,33 @@
+name = "Solarized"
+
+[dark]
+bg = "#002b36"
+dialog_bg = "#073642"
+fg = "#93a1a1"
+accent = "#50aae6"
+accent_secondary = "#8c91dc"
+highlight = "#c8a01e"
+muted = "#879ba5"
+success = "#a0b41e"
+warning = "#e66e3c"
+danger = "#fa6e69"
+border = "#586e75"
+selection_bg = "#073642"
+selection_fg = "#aab9b9"
+graph_line = "#4bbeb4"
+
+[light]
+bg = "#fdf6e3"
+dialog_bg = "#eee8d5"
+fg = "#465a64"
+accent = "#145f96"
+accent_secondary = "#555aa0"
+highlight = "#8c6900"
+muted = "#647378"
+success = "#647300"
+warning = "#a53c0f"
+danger = "#b42828"
+border = "#93a1a1"
+selection_bg = "#dcd7c3"
+selection_fg = "#324650"
+graph_line = "#1e827d"
