+++ crates/theme/src/themes/tokyo-night.toml
@@ -1,0 +1,33 @@
+name = "Tokyo Night"
+
+[dark]
+bg = "#1a1b26"
+dialog_bg = "#24283b"
+fg = "#c0caf5"
+accent = "#7aa2f7"
+accent_secondary = "#bb9af7"
+highlight = "#e0af68"
+muted = "#7882aa"
+success = "#9ece6a"
+warning = "#ff9e64"
+danger = "#f7768e"
+border = "#292e42"
+selection_bg = "#3b4261"
+selection_fg = "#c0caf5"
+graph_line = "#7dcfff"
+
+[light]
+bg = "#e1e2e7"
+dialog_bg = "#ffffff"
+fg = "#376091"
+accent = "#1e5fb4"
+accent_secondary = "#783cc3"
+highlight = "#785514"
+muted = "#505873"
+success = "#286432"
+warning = "#825023"
+danger = "#a5374b"
+border = "#b8bbd4"
+selection_bg = "#9ba5c8"
+selection_fg = "#193264"
+graph_line = "#00648c"
